+++ test/libsolidity/semanticTests/viaYul/dirty_calldata_struct.sol
@@ -4,7 +4,7 @@ contract C {
         uint16[] m;
     }
     function f(S calldata s) public pure returns (bool correct) {
-        int8 x = int8(s.m[0]);
+        int8 x = int8(int16(s.m[0]));
         uint r;
         assembly {
             r := x
